Update the DDK interface and add support for gralloc1.
This change updates the headers imported from the Imagination
PowerVR Rogue DDK to the latest versions. The new headers are
fully backwards compatible with the older GPU DDKs.

Support for gralloc1 (v1) has also been added. This is
implemented as runtime detection of the HAL version, rather
than a compile flag, to enable the GPU DDK to be swapped out
without a full platform rebuild. The plan is that the runtime
checks will be removed in a subsequent change once the import
of DDK 1.8 has stabilized.

The headers updated by this change are used by other projects
so this change will break the build unless it is merged in
tandem with other changes.

Test: Manual, youtube video worked
Bug: 34625842

Change-Id: I700bc8845674135ff7a2b137c04a07c9945e28f7